refactor: 오브제 목록 조회 페이징 처리 관련 리팩토링 #255
